requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
apturl==0.5.2
asgiref==3.7.2
async-timeout==4.0.3
attrs==21.2.0
bcrypt==3.2.0
blinker==1.4
Brlapi==0.8.3
CacheControl==0.12.10
cachy==0.3.0
certifi==2020.6.20
chardet==4.0.0
cleo==0.8.1
click==8.0.3
clikit==0.6.2
colorama==0.4.4
command-not-found==0.3
crashtest==0.3.1
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
distlib==0.3.4
distro==1.7.0
distro-info==1.1+ubuntu0.1
Django==4.2.7
djangorestframework==3.14.0
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
duplicity==0.8.21
fasteners==0.14.1
filelock==3.6.0
frozenlist==1.4.1
future==0.18.2
html5lib==1.1
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
jeepney==0.7.1
jsonschema==3.2.0
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.0.1
monotonic==1.6
more-itertools==8.10.0
msgpack==1.0.3
multidict==6.0.4
netifaces==0.11.0
numpy==1.26.2
oauthlib==3.2.0
olefile==0.46
packaging==21.3
pandas==2.1.3
paramiko==2.9.3
pastel==0.2.1
pexpect==4.8.0
Pikabu==0.0.3
pikabu_api==0.0.1
Pillow==9.0.1
pkginfo==1.8.2
platformdirs==2.5.1
poetry==1.1.12
poetry-core==1.0.7
protobuf==3.12.4
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pycairo==1.20.1
pycups==2.0.1
pydantic==2.5.2
pydantic_core==2.14.5
PyGObject==3.42.1
PyJWT==2.3.0
pylev==1.2.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
pyrsistent==0.18.1
pyTelegramBotAPI==4.14.1
python-apt==2.4.0+ubuntu2
python-box==7.1.1
python-dateutil==2.8.2
python-debian==0.1.43+ubuntu1.1
python-dotenv==0.19.2
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
reportlab==3.6.8
requests==2.25.1
requests-toolbelt==0.9.1
SecretStorage==3.3.1
shellingham==1.4.0
six==1.16.0
sqlparse==0.4.4
systemd-python==234
texttable==1.6.4
tomlkit==0.9.2
typing_extensions==4.8.0
tzdata==2023.3
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
usb-creator==0.3.7
virtualenv==20.13.0+ds
voluptuous==0.14.1
wadllib==1.3.6
webencodings==0.5.1
websocket-client==1.2.3
xdg==5
xkit==0.0.0
yarl==1.9.4
zipp==1.0.0
